requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
jax==0.4.26
jaxlib==0.4.26
Jinja2==3.1.3
lightning==2.2.1
lightning-utilities==0.11.2
lmdb==1.4.1
MarkupSafe==2.1.5
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
protobuf==4.25.3
psutil==5.9.8
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.13.0
sentry-sdk==1.44.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
torch==2.2.2
torchmetrics==1.3.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.6
yacs==0.1.8
yarl==1.9.4
